Added support for alliance logos
Added support for alliance logos on the "We <3 [team]" image. For a list of the alliance logos as well as instructions on what to do in between matches, check out AllianceLogoCollection.ino. Still some issues with randomization outside of main.ino, and there is a problem with the bling board resetting. The latter may just be a power supply problem, however.